backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-CORS==3.0.10
flask-socketio==5.1.1
python-dotenv==0.19.1